Remove UpdateProfileForm component and integrate profile editing functionality into UserProfileInfo.
